backend/requirements.txt
# Vestora Backend — Python Dependencies
fastapi==0.115.6
uvicorn[standard]==0.34.0
motor==3.6.0
pymongo>=4.9,<4.10
pydantic[email]==2.10.3
pydantic-settings==2.7.0
python-jose[cryptography]==3.3.0
bcrypt>=4.1.0
python-multipart==0.0.19
httpx==0.28.1
python-dotenv==1.0.1
Pillow==11.1.0
pillow-heif>=1.2.0
pillow-avif-plugin>=1.4.0
boto3>=1.35.0
google-auth>=2.29.0
requests>=2.31.0
aiosmtplib>=3.0.0
